dev: rework PCI to add type1 header
This adds the possiblity to have PCI device with type 1 config header
(PCI-to-PCI Bridge).

To achieve this, two classes are added PciEndpoint and PciBridge which
inherit the current PciDevice:

- PciDevice acts now on the common part of the configuration (based on
  PCIe 5.0 specification, the first 16 bytes + capability pointer +
  interrupt line/pin). This should not be inherited directly.
- PciEndpoint replace PciDevice for all devices connected as an
  endpoint (GPU, network card, ...) and use the type 0 config header.
- PciBridge add the possibility for PCI bridges using the type 1 config
  header.

The PXCAP union for PCIe capabilities is also updated with all
registers and with the correct size.

This commit is a step into the PCIe architecture implementation from
the following issue: https://github.com/gem5/gem5/issues/1862